board-n1: Code Cleanup fix all warnings
Fixed following Warnings:
arch/arm/mach-tegra/board-n1.c:1148: warning: initialization makes integer from pointer without a cast
arch/arm/mach-tegra/board-n1.c:1150: warning: initialization makes integer from pointer without a cast
arch/arm/mach-tegra/board-n1.c: In function 'n1_touch_exit_hw':
arch/arm/mach-tegra/board-n1.c:1342: warning: comparison between pointer and integer
arch/arm/mach-tegra/board-n1.c:1344: warning: comparison between pointer and integer
arch/arm/mach-tegra/board-n1.c:1346: warning: comparison between pointer and integer
arch/arm/mach-tegra/board-n1.c:1351: warning: comparison between pointer and integer
arch/arm/mach-tegra/board-n1.c:1353: warning: comparison between pointer and integer
arch/arm/mach-tegra/board-n1.c:1355: warning: comparison between pointer and integer
arch/arm/mach-tegra/board-n1.c: At top level:
arch/arm/mach-tegra/board-n1.c:1407: warning: initialization from incompatible pointer type
arch/arm/mach-tegra/board-n1.c:1408: warning: initialization from incompatible pointer type
arch/arm/mach-tegra/board-n1.c:1409: warning: initialization from incompatible pointer type
arch/arm/mach-tegra/board-n1.c:1410: warning: initialization from incompatible pointer type
arch/arm/mach-tegra/board-n1.c:1413: warning: initialization makes integer from pointer without a cast
arch/arm/mach-tegra/board-n1.c:1414: warning: initialization makes integer from pointer without a cast
arch/arm/mach-tegra/board-n1.c:1415: warning: initialization makes integer from pointer without a cast
arch/arm/mach-tegra/board-n1.c:1416: warning: initialization makes integer from pointer without a cast
arch/arm/mach-tegra/board-n1.c:1434: warning: initialization makes integer from pointer without a cast
arch/arm/mach-tegra/board-n1.c:1622: warning: large integer implicitly truncated to unsigned type
arch/arm/mach-tegra/board-n1.c:1628: warning: large integer implicitly truncated to unsigned type
arch/arm/mach-tegra/board-n1.c:1632: warning: function declaration isn't a prototype
arch/arm/mach-tegra/board-n1.c:1699: warning: initialization from incompatible pointer type
arch/arm/mach-tegra/board-n1.c:1700: warning: initialization from incompatible pointer type
arch/arm/mach-tegra/board-n1.c:1767: warning: large integer implicitly truncated to unsigned type
arch/arm/mach-tegra/board-n1.c:1779: warning: large integer implicitly truncated to unsigned type

Fuck you Samsung.

+++ include/media/s5k6aafx.h
@@ -184,7 +184,7 @@ typedef enum {
 
 
 struct s5k6aafx_reg_8 {
-	u8 addr;
+	u16 addr;
 	u8 val;
 };
 struct s5k6aafx_reg {
